umbra: Initialise something that works
Fixes a whole lotta build errors. Might want to fix the hack for
imx179.c later, but at least it compiles.

* Fix the stupid include brackets, the header files are literally
there, use quotation and not brackets for that
* Fix the variable 'type' breaking build due to aggressive loop
optimization being enabled, this may not work and may break some stuff
* Get a proper .gitignore

There's a built firmware in this, I'll add it to gitignore list because
they're built alongside the kernel